refactor(langgraph): 서브그래프 분해 및 중복 헬퍼 통합
runtime.py(4437줄, 단일 클래스)가 그래프 배선/6개 서브그래프 노드/실험적
프로파일/실행 경계를 전부 담고 있던 God Object 문제와, application/workflows
6개 파일에 거의 동일한 스키마 검증 헬퍼가 복붙돼 있던 DRY 위반을 해소한다.

- adapters/langgraph/graph_state.py 신설: GraphState/ParentGraphState 타입,
  키 상수, request_from_state 등 순수 헬퍼 분리
- adapters/langgraph/subgraphs/ 패키지 신설: SIX_ROLE_BASELINE 6개
  서브그래프(request_understanding/acquisition/context_retrieval/
  work_analysis/planning/review)와 THREE_STAGE/SINGLE_BASELINE 실험적
  프로파일을 각각 독립 클래스로 분해, 명시적 의존성 주입으로 결합도 감소
- application/workflows/_schema_support.py 신설: 6개 파일에 중복 정의돼
  있던 검증 헬퍼 12종을 error_cls 매개변수화된 공용 구현으로 통합하고
  각 파일은 functools.partial로 재바인딩(호출부 무변경, 예외 타입 보존)
- runtime.py에서 어떤 그래프에도 연결되지 않던 dead code(미사용 노드
  함수 6개, _node_handler의 죽은 매핑 키 4개) 제거
- runtime.py 4437 -> 2146줄(52% 감소), WorkflowRuntime 포트 시그니처 및
  테스트가 직접 참조하는 준-공개 속성(_graph/_topology/_node_handler 등)
  전부 보존, 순수 구조 이동만 수행(동작 변경 없음)

검증: pytest -q(전체 1108개) 통과, ruff check/format 통과,
mypy src/google_work_agent(162개 파일) 통과
