+++ src/google_work_agent/adapters/langgraph/runtime.py
@@ -9,18 +9,42 @@ from hashlib import sha256
 from json import dumps, loads
 from pathlib import Path
 from threading import Lock
-from typing import Any, Final, Literal, NotRequired, TypedDict, cast
+from typing import Any, cast
 
 from langgraph.checkpoint.sqlite import SqliteSaver
 from langgraph.graph import END, START, StateGraph
 from langgraph.types import Command, interrupt
 
-from google_work_agent.adapters.langgraph.agent_kernel import (
-    build_agent_local_state,
-    merge_trace_context,
-    record_llm_result,
+from google_work_agent.adapters.langgraph.graph_state import (
+    GraphState,
+    _acquired_resource_by_handle,
+    _require_state_value,
+    _resource_handle_for_ref,
+    _stored_resource_type_for_acquired_resource,
+    request_from_state,
 )
 from google_work_agent.adapters.langgraph.profiles import GraphProfile
+from google_work_agent.adapters.langgraph.subgraphs.acquisition import AcquisitionSubgraph
+from google_work_agent.adapters.langgraph.subgraphs.context_retrieval import (
+    ContextRetrieverSubgraph,
+)
+from google_work_agent.adapters.langgraph.subgraphs.planning import (
+    PlanningSubgraph,
+    planning_mode_from_request_intent,
+)
+from google_work_agent.adapters.langgraph.subgraphs.request_understanding import (
+    RequestUnderstandingSubgraph,
+)
+from google_work_agent.adapters.langgraph.subgraphs.review import ReviewSubgraph
+from google_work_agent.adapters.langgraph.subgraphs.single_workflow import (
+    SingleWorkflowSubgraph,
+)
+from google_work_agent.adapters.langgraph.subgraphs.three_stage import (
+    ThreeStageOneSubgraph,
+    ThreeStageReviewSubgraph,
+    ThreeStageTwoSubgraph,
+)
+from google_work_agent.adapters.langgraph.subgraphs.work_analysis import WorkAnalysisSubgraph
 from google_work_agent.application import (
     BlockRunCommand,
     BlockRunService,
@@ -77,7 +101,6 @@ from google_work_agent.application.calendar_conflicts import (
     evidence_calendar_conflict_risk,
 )
 from google_work_agent.application.feasibility import evidence_feasibility_risk
-from google_work_agent.application.observability import ObservabilityContext
 from google_work_agent.application.task_duplicates import (
     TASK_CREATE_TOOL,
     evidence_duplicate_risk,
@@ -85,55 +108,31 @@ from google_work_agent.application.task_duplicates import (
 from google_work_agent.application.workflows import (
     AcquisitionResultV1,
     ActionPlanDraftV1,
-    AdditionalAcquisitionRequestV1,
-    AgentLocalStateV1,
-    AnswerDraftV1,
     ApiDiscoveryAcquisitionAgent,
-    ContextBundleV1,
     ContextRetrievalAgent,
-    ContextRetrievalResultV1,
     DomainValidationResult,
     DomainValidationService,
-    EvidenceDraftV1,
-    EvidenceSelectionOutputV1,
     GraphStateUpdateV1,
     MultiAgentGraphState,
     PlanReviewAgent,
     PlanReviewResultV1,
     RequestIntentV1,
     RequestUnderstandingAgent,
-    RequestUnderstandingOutputV1,
     ReviewResult,
-    RunBudgetV1,
     SolutionPlanningAgent,
-    SourceFetchPlanV1,
-    SourcePlanningOutputV1,
-    SufficiencyOutputV1,
     SupervisorDecisionV1,
     SupervisorTarget,
     WorkAnalysisAgent,
-    WorkAnalysisResultV1,
     WorkflowPhase,
     route_supervisor,
-    validate_acquisition_result_v1,
-    validate_action_plan_draft_v1,
-    validate_answer_draft_v1,
-    validate_context_retrieval_result_v1,
 )
 from google_work_agent.application.workflows.profile_fused import (
-    PROFILE_FUSED_PLANNING_OUTPUT_SCHEMA,
-    PROFILE_REQUEST_SOURCE_OUTPUT_SCHEMA,
-    ProfilePlanningProjectionV1,
-    ProfileReasonPlanOutputV1,
-    ProfileRequestSourceOutputV1,
     load_profile_single_reason_plan_prompt_reference,
     load_profile_single_request_source_prompt_reference,
     load_profile_single_self_review_prompt_reference,
     load_profile_single_self_review_recheck_prompt_reference,
     load_profile_three_stage1_prompt_reference,
     load_profile_three_stage2_prompt_reference,
-    validate_profile_reason_plan_output_v1,
-    validate_profile_request_source_output_v1,
 )
 from google_work_agent.application.write_actions import (
     ClaimWriteActionCommand,
@@ -158,7 +157,6 @@ from google_work_agent.ports import (
     PromptReference,
     ResourceRefRecord,
     ResourceSource,
-    StoredResourceType,
     UnitOfWork,
     WorkflowCancelRequest,
     WorkflowInvocationResult,
@@ -171,149 +169,6 @@ from google_work_agent.ports import (
 from google_work_agent.ports.repositories import ActionRecord
 
 JsonObject = dict[str, object]
-
-
-class ParentGraphState(MultiAgentGraphState):
-    """State projected from a native subgraph back to the parent graph."""
-
-    __request__: WorkflowStartRequest
-    __target__: str
-    __logical_target__: str
-    __modify_review_plan_id__: NotRequired[str | None]
-    __modify_review_version__: NotRequired[int | None]
-    __modify_review_risks__: NotRequired[dict[str, dict[str, object]] | None]
-    __replan_from_plan_id__: NotRequired[str]
-
-
-class GraphState(TypedDict):
-    """Executable graph state, including invocation-local subgraph projections."""
-
-    schema_version: Literal[1]
-    run_id: str
-    conversation_id: str
-    thread_id: str
-    workflow_phase: str
-    request_intent: RequestIntentV1 | None
-    source_fetch_plans: list[SourceFetchPlanV1]
-    acquisition_result: AcquisitionResultV1 | None
-    context_result: ContextRetrievalResultV1 | None
-    analysis_result: WorkAnalysisResultV1 | None
-    answer_draft: AnswerDraftV1 | None
-    plan_draft: ActionPlanDraftV1 | None
-    plan_review: PlanReviewResultV1 | None
-    approved_plan_id: str | None
-    execution_summary: dict[str, object] | None
-    verification_summary: dict[str, object] | None
-    finalize_intent: object | None
-    user_interrupt: object | None
-    retry_budget: RunBudgetV1
-    prompt_context: dict[str, object]
-    trace_context: dict[str, object]
-    context_bundle: NotRequired[ContextBundleV1]
-    evidence_drafts: NotRequired[list[EvidenceDraftV1]]
-    llm_provider_result: NotRequired[dict[str, object] | None]
-    __request__: WorkflowStartRequest
-    __target__: str
-    __logical_target__: str
-    __modify_review_plan_id__: NotRequired[str | None]
-    __modify_review_version__: NotRequired[int | None]
-    __modify_review_risks__: NotRequired[dict[str, dict[str, object]] | None]
-    __replan_from_plan_id__: NotRequired[str]
-    __request_agent_local__: NotRequired[AgentLocalStateV1]
-    __request_output__: NotRequired[RequestUnderstandingOutputV1]
-    __acquisition_agent_local__: NotRequired[AgentLocalStateV1]
-    __acquisition_planning_output__: NotRequired[SourcePlanningOutputV1]
-    __context_agent_local__: NotRequired[AgentLocalStateV1]
-    __context_selection_output__: NotRequired[EvidenceSelectionOutputV1]
-    __context_sufficiency_output__: NotRequired[SufficiencyOutputV1]
-    __analysis_agent_local__: NotRequired[AgentLocalStateV1]
-    __planning_agent_local__: NotRequired[AgentLocalStateV1]
-    __planning_mode__: NotRequired[str]
-    __planning_result__: NotRequired[AnswerDraftV1 | ActionPlanDraftV1]
-    __review_agent_local__: NotRequired[AgentLocalStateV1]
-    __review_mode__: NotRequired[str]
-    __profile_agent_local__: NotRequired[AgentLocalStateV1]
-    __profile_request_source_output__: NotRequired[ProfileRequestSourceOutputV1]
-    __profile_reason_plan_output__: NotRequired[ProfileReasonPlanOutputV1]
-
-
-REQUEST_AGENT_LOCAL_KEY: Final = "__request_agent_local__"
-REQUEST_OUTPUT_KEY: Final = "__request_output__"
-ACQUISITION_AGENT_LOCAL_KEY: Final = "__acquisition_agent_local__"
-ACQUISITION_PLANNING_OUTPUT_KEY: Final = "__acquisition_planning_output__"
-CONTEXT_AGENT_LOCAL_KEY: Final = "__context_agent_local__"
-CONTEXT_SELECTION_OUTPUT_KEY: Final = "__context_selection_output__"
-CONTEXT_SUFFICIENCY_OUTPUT_KEY: Final = "__context_sufficiency_output__"
-ANALYSIS_AGENT_LOCAL_KEY: Final = "__analysis_agent_local__"
-PLANNING_AGENT_LOCAL_KEY: Final = "__planning_agent_local__"
-PLANNING_MODE_KEY: Final = "__planning_mode__"
-REVIEW_AGENT_LOCAL_KEY: Final = "__review_agent_local__"
-REVIEW_MODE_KEY: Final = "__review_mode__"
-PROFILE_AGENT_LOCAL_KEY: Final = "__profile_agent_local__"
-PROFILE_REQUEST_SOURCE_OUTPUT_KEY: Final = "__profile_request_source_output__"
-PROFILE_REASON_PLAN_OUTPUT_KEY: Final = "__profile_reason_plan_output__"
-
-
-def _require_state_value[StateValueT](
-    value: StateValueT | None,
-    field_name: str,
-) -> StateValueT:
-    if value is None:
-        raise ValueError(f"graph state is missing required field: {field_name}")
-    return value
-
-
-def _resource_handle_for_ref(resource_ref: ResourceRefRecord) -> str:
-    prefixes = {
-        ("GMAIL", "THREAD"): "gmail_thread",
-        ("GMAIL", "MESSAGE"): "gmail_message",
-        ("TASKS", "TASK_LIST"): "task_list",
-        ("TASKS", "TASK"): "task",
-        ("CALENDAR", "CALENDAR"): "calendar",
-        ("CALENDAR", "EVENT"): "calendar_event",
-    }
-    prefix = prefixes.get((resource_ref.source.value, resource_ref.resource_type.value))
-    if prefix is None:
-        raise LookupError(f"unsupported persisted resource reference: {resource_ref.id}")
-    return f"{prefix}:{resource_ref.resource_id}"
-
-
-def _acquired_resource_by_handle(
-    *, acquisition_result: AcquisitionResultV1, resource_handle: str
-) -> dict[str, object] | None:
-    source_summaries = acquisition_result["source_summaries"]
-    for summary in source_summaries:
-        source = summary.get("source")
-        resources = summary.get("resources")
-        if not isinstance(source, str) or not isinstance(resources, list):
-            continue
-        for resource in resources:
-            if not isinstance(resource, dict):
-                continue
-            if resource.get("resource_handle") != resource_handle:
-                continue
-            payload = resource.get("payload")
-            if not isinstance(payload, dict):
-                raise LookupError(f"acquired resource payload is invalid: {resource_handle}")
-            return {**resource, "source": source, "payload": payload}
-    return None
-
-
-def _stored_resource_type_for_acquired_resource(
-    *, source: ResourceSource, resource_type: str
-) -> StoredResourceType:
-    mapping = {
-        (ResourceSource.GMAIL, "gmail_thread"): StoredResourceType.THREAD,
-        (ResourceSource.GMAIL, "gmail_message"): StoredResourceType.MESSAGE,
-        (ResourceSource.TASKS, "task_list"): StoredResourceType.TASK_LIST,
-        (ResourceSource.TASKS, "task"): StoredResourceType.TASK,
-        (ResourceSource.CALENDAR, "calendar"): StoredResourceType.CALENDAR,
-        (ResourceSource.CALENDAR, "calendar_event"): StoredResourceType.EVENT,
-    }
-    stored_type = mapping.get((source, resource_type))
-    if stored_type is None:
-        raise LookupError(f"unsupported acquired resource type: {source.value}/{resource_type}")
-    return stored_type
 
 
 class LangGraphWorkflowRuntime(WorkflowRuntime):
@@ -533,22 +388,92 @@ class LangGraphWorkflowRuntime(WorkflowRuntime):
             now_ms=now_ms,
             gateway=gateway,
         )
-        self._request_subgraph = self._build_request_subgraph()
-        self._acquisition_subgraph = self._build_acquisition_subgraph()
-        self._context_subgraph = self._build_context_subgraph()
-        self._analysis_subgraph = self._build_analysis_subgraph()
-        self._planning_subgraph = self._build_planning_subgraph()
-        self._review_subgraph = self._build_review_subgraph()
+        self._request_subgraph = RequestUnderstandingSubgraph(
+            agent=self._request_understanding,
+            id_factory=id_factory,
+            graph_profile=self._graph_profile,
+            transition_run=self._transition_run,
+            merge_decision=self._merge_decision,
+        ).build()
+        self._acquisition_subgraph = AcquisitionSubgraph(
+            agent=self._acquisition,
+            id_factory=id_factory,
+            graph_profile=self._graph_profile,
+            transition_run=self._transition_run,
+            merge_decision=self._merge_decision,
+        ).build()
+        self._context_subgraph = ContextRetrieverSubgraph(
+            agent=self._context,
+            id_factory=id_factory,
+            graph_profile=self._graph_profile,
+            merge_decision=self._merge_decision,
+        ).build()
+        self._analysis_subgraph = WorkAnalysisSubgraph(
+            agent=self._analysis,
+            id_factory=id_factory,
+            graph_profile=self._graph_profile,
+            transition_run=self._transition_run,
+            merge_decision=self._merge_decision,
+        ).build()
+        self._planning_subgraph = PlanningSubgraph(
+            agent=self._planning,
+            id_factory=id_factory,
+            graph_profile=self._graph_profile,
+            merge_decision=self._merge_decision,
+        ).build()
+        self._review_subgraph = ReviewSubgraph(
+            agent=self._review,
+            id_factory=id_factory,
+            graph_profile=self._graph_profile,
+            merge_decision=self._merge_decision,
+        ).build()
         self._three_stage_one_subgraph: Any = None
         self._three_stage_two_subgraph: Any = None
         self._three_stage_review_subgraph: Any = None
         if self._graph_profile is GraphProfile.THREE_STAGE:
-            self._three_stage_one_subgraph = self._build_three_stage_one_subgraph()
-            self._three_stage_two_subgraph = self._build_three_stage_two_subgraph()
-            self._three_stage_review_subgraph = self._build_three_stage_review_subgraph()
+            assert self._three_stage1_prompt_ref is not None
+            assert self._three_stage2_prompt_ref is not None
+            self._three_stage_one_subgraph = ThreeStageOneSubgraph(
+                request_understanding_agent=self._request_understanding,
+                acquisition_agent=self._acquisition,
+                prompt_ref=self._three_stage1_prompt_ref,
+                id_factory=id_factory,
+                graph_profile=self._graph_profile,
+                transition_run=self._transition_run,
+                merge_decision=self._merge_decision,
+            ).build()
+            self._three_stage_two_subgraph = ThreeStageTwoSubgraph(
+                request_understanding_agent=self._request_understanding,
+                planning_agent=self._planning,
+                prompt_ref=self._three_stage2_prompt_ref,
+                id_factory=id_factory,
+                graph_profile=self._graph_profile,
+                transition_run=self._transition_run,
+                merge_decision=self._merge_decision,
+            ).build()
+            self._three_stage_review_subgraph = ThreeStageReviewSubgraph(
+                agent=self._review,
+                id_factory=id_factory,
+                graph_profile=self._graph_profile,
+                merge_decision=self._merge_decision,
+            ).build()
         self._single_workflow_subgraph: Any = None
         if self._graph_profile is GraphProfile.SINGLE_BASELINE:
-            self._single_workflow_subgraph = self._build_single_workflow_subgraph()
+            assert self._single_request_source_prompt_ref is not None
+            assert self._single_reason_plan_prompt_ref is not None
+            assert self._single_review is not None
+            self._single_workflow_subgraph = SingleWorkflowSubgraph(
+                request_understanding_agent=self._request_understanding,
+                acquisition_agent=self._acquisition,
+                planning_agent=self._planning,
+                review_agent=self._single_review,
+                request_source_prompt_ref=self._single_request_source_prompt_ref,
+                reason_plan_prompt_ref=self._single_reason_plan_prompt_ref,
+                id_factory=id_factory,
+                graph_profile=self._graph_profile,
+                transition_run=self._transition_run,
+                merge_decision=self._merge_decision,
+            ).build()
         self._native_agent_subgraphs = self._native_subgraphs_for_profile()
         self._topology = self._topology_for_profile()
         self._graph = self._build_graph()
@@ -748,10 +673,6 @@ class LangGraphWorkflowRuntime(WorkflowRuntime):
             "planning": self._planning_subgraph,
             "review": self._review_subgraph,
             "single_workflow": self._single_workflow_subgraph,
-            "source_planning": self._source_planning_node,
-            "api_acquisition": self._api_acquisition_node,
-            "context_retrieval": self._context_retrieval_node,
-            "plan_review": self._plan_review_node,
             "domain_validation": self._domain_validation_node,
             "stage_one": self._three_stage_one_subgraph,
             "stage_two": self._three_stage_two_subgraph,
@@ -776,2165 +697,6 @@ class LangGraphWorkflowRuntime(WorkflowRuntime):
                 "stage_three": self._three_stage_review_subgraph,
             }
         return {"single_workflow": self._single_workflow_subgraph}
-
-    def _build_request_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._request_subgraph_init_node)
-        graph.add_node("classify", self._request_subgraph_classify_node)
-        graph.add_node("finalize", self._request_subgraph_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "classify")
-        graph.add_edge("classify", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="request_understanding_subgraph")
-
-    def _build_acquisition_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._acquisition_subgraph_init_node)
-        graph.add_node("plan_sources", self._acquisition_subgraph_plan_sources_node)
-        graph.add_node("plan_validate", self._acquisition_subgraph_plan_validate_node)
-        graph.add_node("deterministic_read", self._acquisition_subgraph_read_node)
-        graph.add_node("result_validate", self._acquisition_subgraph_result_validate_node)
-        graph.add_node("finalize", self._acquisition_subgraph_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "plan_sources")
-        graph.add_edge("plan_sources", "plan_validate")
-        graph.add_conditional_edges(
-            "plan_validate",
-            self._route_acquisition_plan_validate,
-            {
-                "deterministic_read": "deterministic_read",
-                "finalize": "finalize",
-            },
-        )
-        graph.add_edge("deterministic_read", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="acquisition_subgraph")
-
-    def _request_subgraph_init_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "start_analysis")
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="request_understanding",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={
-                "request_text": request.request_text,
-                "entry_mode": request.entry_mode,
-                "selected_resource_ids": list(request.selected_resource_ids),
-            },
-            prompt_ref=self._request_understanding.prompt_ref,
-        )
-        return {
-            **state,
-            REQUEST_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="request_understanding",
-                agent_role="request_understanding",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="request_understanding",
-                node_name="init",
-                prompt_ref=self._request_understanding.prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _request_subgraph_classify_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[REQUEST_AGENT_LOCAL_KEY])
-        llm_result = self._request_understanding.invoke_classify_llm(request)
-        output = self._request_understanding.build_output_from_llm_result(llm_result)
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "CLASSIFY_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], output)
-        return {
-            **state,
-            REQUEST_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            REQUEST_OUTPUT_KEY: output,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="request_understanding",
-                agent_role="request_understanding",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="request_understanding",
-                node_name="classify",
-                llm_call_id=f"{request.run_id}:request_understanding.classify",
-                prompt_ref=self._request_understanding.prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-            ),
-        }
-
-    def _request_subgraph_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[REQUEST_AGENT_LOCAL_KEY])
-        request = self._request_from_state(state)
-        output = state[REQUEST_OUTPUT_KEY]
-        decision = route_supervisor(
-            phase=WorkflowPhase.REQUEST_ANALYSIS,
-            state=cast(MultiAgentGraphState, state),
-            result=output,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, output["result"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="request_understanding",
-                    agent_role="request_understanding",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="request_understanding",
-                    node_name="finalize",
-                ),
-                REQUEST_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            },
-            self._request_understanding.build_state_update(output, request=request),
-            decision,
-        )
-        merged.pop(REQUEST_AGENT_LOCAL_KEY, None)
-        merged.pop(REQUEST_OUTPUT_KEY, None)
-        return merged
-
-    def _acquisition_subgraph_init_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_retrieval")
-        additional: AdditionalAcquisitionRequestV1 | None = None
-        context_result = state.get("context_result")
-        analysis_result = state.get("analysis_result")
-        if context_result is not None:
-            additional = context_result["additional_acquisition_request"]
-        if additional is None and analysis_result is not None:
-            additional = analysis_result["additional_acquisition_request"]
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="api_discovery_acquisition",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={
-                "request_intent": _require_state_value(state["request_intent"], "request_intent"),
-                "additional_acquisition_request": additional,
-                "entry_mode": request.entry_mode,
-            },
-            prompt_ref=self._acquisition.prompt_ref,
-        )
-        next_state: GraphState = {
-            **state,
-            ACQUISITION_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="acquisition",
-                agent_role="api_discovery_acquisition",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="acquisition",
-                node_name="init",
-                prompt_ref=self._acquisition.prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-        return next_state
-
-    def _acquisition_subgraph_plan_sources_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[ACQUISITION_AGENT_LOCAL_KEY])
-        additional: AdditionalAcquisitionRequestV1 | None = None
-        context_result = state.get("context_result")
-        analysis_result = state.get("analysis_result")
-        if context_result is not None:
-            additional = context_result["additional_acquisition_request"]
-        if additional is None and analysis_result is not None:
-            additional = analysis_result["additional_acquisition_request"]
-        llm_result = self._acquisition.invoke_plan_sources_llm(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            request=request,
-            additional_acquisition_request=additional,
-        )
-        output = self._acquisition.build_planning_output_from_llm_result(llm_result)
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "PLAN_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], output)
-        return {
-            **state,
-            ACQUISITION_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            ACQUISITION_PLANNING_OUTPUT_KEY: output,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="acquisition",
-                agent_role="api_discovery_acquisition",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="acquisition",
-                node_name="plan_sources",
-                llm_call_id=f"{request.run_id}:acquisition.plan_sources",
-                prompt_ref=self._acquisition.prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-            ),
-        }
-
-    def _acquisition_subgraph_plan_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[ACQUISITION_AGENT_LOCAL_KEY])
-        planning_output = state[ACQUISITION_PLANNING_OUTPUT_KEY]
-        source_fetch_plans = planning_output.get("source_fetch_plans")
-        if not isinstance(source_fetch_plans, list):
-            raise TypeError("acquisition planning output is missing source_fetch_plans")
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "PLAN_VALIDATED"
-        updated_local["typed_result"] = planning_output
-        return {
-            **state,
-            ACQUISITION_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="acquisition",
-                agent_role="api_discovery_acquisition",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="acquisition",
-                node_name="plan_validate",
-            ),
-        }
-
-    def _route_acquisition_plan_validate(self, state: GraphState) -> str:
-        planning_output = state[ACQUISITION_PLANNING_OUTPUT_KEY]
-        return "deterministic_read" if planning_output["result"] == "PLAN_READY" else "finalize"
-
-    def _acquisition_subgraph_read_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[ACQUISITION_AGENT_LOCAL_KEY])
-        planning_output = state[ACQUISITION_PLANNING_OUTPUT_KEY]
-        result = self._acquisition.acquire(
-            plans=planning_output["source_fetch_plans"],
-            request=request,
-            request_intent=state.get("request_intent"),
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "READ_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            "acquisition_result": result,
-            ACQUISITION_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="acquisition",
-                agent_role="api_discovery_acquisition",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="acquisition",
-                node_name="deterministic_read",
-            ),
-        }
-
-    def _acquisition_subgraph_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[ACQUISITION_AGENT_LOCAL_KEY])
-        acquisition_result = validate_acquisition_result_v1(state["acquisition_result"])
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = cast(dict[str, object], acquisition_result)
-        return {
-            **state,
-            "acquisition_result": acquisition_result,
-            ACQUISITION_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="acquisition",
-                agent_role="api_discovery_acquisition",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="acquisition",
-                node_name="result_validate",
-            ),
-        }
-
-    def _acquisition_subgraph_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[ACQUISITION_AGENT_LOCAL_KEY])
-        planning_output = state[ACQUISITION_PLANNING_OUTPUT_KEY]
-        current: GraphState = {
-            **state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="acquisition",
-                agent_role="api_discovery_acquisition",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="acquisition",
-                node_name="finalize",
-            ),
-        }
-        if planning_output["result"] != "PLAN_READY":
-            decision = route_supervisor(
-                phase=WorkflowPhase.SOURCE_PLANNING,
-                state=cast(MultiAgentGraphState, current),
-                result=planning_output,
-            )
-            updated_local = dict(local_state)
-            updated_local["node_state"] = "FINALIZED"
-            updated_local["disposition"] = {
-                "schema_version": 1,
-                "status": cast(str, planning_output["result"]),
-                "next_target": cast(str, decision["target"]),
-                "reason_code": cast(str | None, decision.get("reason_code")),
-            }
-            merged = self._merge_decision(
-                {**current, ACQUISITION_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local)},
-                self._acquisition.build_planning_state_update(planning_output),
-                decision,
-            )
-        else:
-            acquisition_result = _require_state_value(
-                state["acquisition_result"], "acquisition_result"
-            )
-            decision = route_supervisor(
-                phase=WorkflowPhase.API_ACQUISITION,
-                state=cast(MultiAgentGraphState, current),
-                result=acquisition_result,
-            )
-            updated_local = dict(local_state)
-            updated_local["node_state"] = "FINALIZED"
-            updated_local["disposition"] = {
-                "schema_version": 1,
-                "status": cast(str, acquisition_result["status"]),
-                "next_target": cast(str, decision["target"]),
-                "reason_code": cast(str | None, decision.get("reason_code")),
-            }
-            merged = self._merge_decision(
-                {**current, ACQUISITION_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local)},
-                {
-                    **self._acquisition.build_planning_state_update(planning_output),
-                    **self._acquisition.build_acquisition_state_update(acquisition_result),
-                },
-                decision,
-            )
-        merged.pop(ACQUISITION_AGENT_LOCAL_KEY, None)
-        merged.pop(ACQUISITION_PLANNING_OUTPUT_KEY, None)
-        return merged
-
-    def _build_context_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._context_subgraph_init_node)
-        graph.add_node("select_evidence", self._context_subgraph_select_evidence_node)
-        graph.add_node("selection_validate", self._context_subgraph_selection_validate_node)
-        graph.add_node("assess_sufficiency", self._context_subgraph_assess_sufficiency_node)
-        graph.add_node("finalize", self._context_subgraph_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "select_evidence")
-        graph.add_edge("select_evidence", "selection_validate")
-        graph.add_edge("selection_validate", "assess_sufficiency")
-        graph.add_edge("assess_sufficiency", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="context_retriever_subgraph")
-
-    def _build_analysis_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._analysis_subgraph_init_node)
-        graph.add_node("analyze", self._analysis_subgraph_analyze_node)
-        graph.add_node("result_validate", self._analysis_subgraph_result_validate_node)
-        graph.add_node("finalize", self._analysis_subgraph_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "analyze")
-        graph.add_edge("analyze", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="work_analysis_subgraph")
-
-    def _build_planning_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._planning_subgraph_init_node)
-        graph.add_node("plan", self._planning_subgraph_plan_node)
-        graph.add_node("result_validate", self._planning_subgraph_result_validate_node)
-        graph.add_node("finalize", self._planning_subgraph_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "plan")
-        graph.add_edge("plan", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="planning_subgraph")
-
-    def _build_review_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._review_subgraph_init_node)
-        graph.add_node("review", self._review_subgraph_review_node)
-        graph.add_node("result_validate", self._review_subgraph_result_validate_node)
-        graph.add_node("finalize", self._review_subgraph_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "review")
-        graph.add_edge("review", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="review_subgraph")
-
-    def _build_three_stage_one_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._three_stage_one_init_node)
-        graph.add_node("request_source", self._three_stage_one_request_source_node)
-        graph.add_node("plan_validate", self._three_stage_one_plan_validate_node)
-        graph.add_node("deterministic_read", self._three_stage_one_read_node)
-        graph.add_node("result_validate", self._three_stage_one_result_validate_node)
-        graph.add_node("finalize", self._three_stage_one_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "request_source")
-        graph.add_edge("request_source", "plan_validate")
-        graph.add_conditional_edges(
-            "plan_validate",
-            self._route_profile_stage_one_plan_validate,
-            {
-                "deterministic_read": "deterministic_read",
-                "finalize": "finalize",
-            },
-        )
-        graph.add_edge("deterministic_read", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="three_stage_one_subgraph")
-
-    def _build_three_stage_two_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._three_stage_two_init_node)
-        graph.add_node("reason_plan", self._three_stage_two_reason_plan_node)
-        graph.add_node("result_validate", self._three_stage_two_result_validate_node)
-        graph.add_node("finalize", self._three_stage_two_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "reason_plan")
-        graph.add_edge("reason_plan", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="three_stage_two_subgraph")
-
-    def _build_three_stage_review_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._three_stage_review_init_node)
-        graph.add_node("review", self._three_stage_review_node)
-        graph.add_node("result_validate", self._three_stage_review_result_validate_node)
-        graph.add_node("finalize", self._three_stage_review_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "review")
-        graph.add_edge("review", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="three_stage_review_subgraph")
-
-    def _build_single_workflow_subgraph(self) -> Any:
-        graph = StateGraph(GraphState, output_schema=ParentGraphState)
-        graph.add_node("init", self._single_workflow_init_node)
-        graph.add_node("request_source", self._single_workflow_request_source_node)
-        graph.add_node("plan_validate", self._single_workflow_plan_validate_node)
-        graph.add_node("deterministic_read", self._single_workflow_read_node)
-        graph.add_node("reason_plan", self._single_workflow_reason_plan_node)
-        graph.add_node("self_review", self._single_workflow_self_review_node)
-        graph.add_node("result_validate", self._single_workflow_result_validate_node)
-        graph.add_node("finalize", self._single_workflow_finalize_node)
-        graph.add_edge(START, "init")
-        graph.add_edge("init", "request_source")
-        graph.add_edge("request_source", "plan_validate")
-        graph.add_conditional_edges(
-            "plan_validate",
-            self._route_single_workflow_plan_validate,
-            {
-                "deterministic_read": "deterministic_read",
-                "reason_plan": "reason_plan",
-                "finalize": "finalize",
-            },
-        )
-        graph.add_edge("deterministic_read", "reason_plan")
-        graph.add_edge("reason_plan", "self_review")
-        graph.add_edge("self_review", "result_validate")
-        graph.add_edge("result_validate", "finalize")
-        graph.add_edge("finalize", END)
-        return graph.compile(name="single_workflow_subgraph")
-
-    def _context_subgraph_init_node(self, state: GraphState) -> GraphState:
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="context_retriever",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={
-                "request_intent": _require_state_value(state["request_intent"], "request_intent"),
-                "acquisition_result": _require_state_value(
-                    state["acquisition_result"], "acquisition_result"
-                ),
-            },
-            prompt_ref=self._context.select_prompt_ref,
-        )
-        return {
-            **state,
-            CONTEXT_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="context_retriever",
-                agent_role="context_retriever",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="context",
-                node_name="init",
-                prompt_ref=self._context.select_prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _context_subgraph_select_evidence_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[CONTEXT_AGENT_LOCAL_KEY])
-        acquisition_result = _require_state_value(state["acquisition_result"], "acquisition_result")
-        segments = self._context.build_segments_from_acquisition(acquisition_result)
-        selection = self._context.select_evidence(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            acquisition_result=acquisition_result,
-            request=request,
-            segments=cast(list[Any], segments),
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "SELECT_EVIDENCE_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], selection)
-        return {
-            **state,
-            CONTEXT_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            CONTEXT_SELECTION_OUTPUT_KEY: selection,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="context_retriever",
-                agent_role="context_retriever",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="context",
-                node_name="select_evidence",
-                llm_call_id=f"{request.run_id}:context.select_evidence",
-                prompt_ref=self._context.select_prompt_ref,
-                llm_call_increment=1,
-            ),
-        }
-
-    def _context_subgraph_selection_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[CONTEXT_AGENT_LOCAL_KEY])
-        selection = state[CONTEXT_SELECTION_OUTPUT_KEY]
-        acquisition_result = _require_state_value(state["acquisition_result"], "acquisition_result")
-        draft_bundle, evidence_drafts = self._context.build_draft_context_bundle(
-            selection_result=selection,
-            acquisition_result=acquisition_result,
-            missing_information=selection["missing_information"],
-            ambiguity=selection["ambiguity"],
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "SELECTION_VALIDATED"
-        updated_local["typed_result"] = cast(dict[str, object], selection)
-        return {
-            **state,
-            CONTEXT_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "context_bundle": draft_bundle,
-            "evidence_drafts": evidence_drafts,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="context_retriever",
-                agent_role="context_retriever",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="context",
-                node_name="selection_validate",
-            ),
-        }
-
-    def _context_subgraph_assess_sufficiency_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[CONTEXT_AGENT_LOCAL_KEY])
-        sufficiency_result, llm_provider_result = self._context.assess_sufficiency(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            acquisition_result=_require_state_value(
-                state["acquisition_result"], "acquisition_result"
-            ),
-            request=request,
-            context_bundle=state["context_bundle"],
-            evidence_drafts=state["evidence_drafts"],
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "SUFFICIENCY_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], sufficiency_result)
-        return {
-            **state,
-            CONTEXT_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            CONTEXT_SUFFICIENCY_OUTPUT_KEY: sufficiency_result,
-            "llm_provider_result": llm_provider_result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="context_retriever",
-                agent_role="context_retriever",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="context",
-                node_name="assess_sufficiency",
-                llm_call_id=f"{request.run_id}:context.assess_sufficiency",
-                prompt_ref=self._context.sufficiency_prompt_ref,
-                llm_call_increment=1,
-            ),
-        }
-
-    def _context_subgraph_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[CONTEXT_AGENT_LOCAL_KEY])
-        selection = state[CONTEXT_SELECTION_OUTPUT_KEY]
-        sufficiency = state[CONTEXT_SUFFICIENCY_OUTPUT_KEY]
-        llm_provider_result = _require_state_value(
-            state.get("llm_provider_result"), "llm_provider_result"
-        )
-        result = validate_context_retrieval_result_v1(
-            self._context.build_result_from_outputs(
-                selection_result=selection,
-                sufficiency_result=sufficiency,
-                acquisition_result=_require_state_value(
-                    state["acquisition_result"], "acquisition_result"
-                ),
-                llm_provider_result=llm_provider_result,
-            )
-        )
-        decision = route_supervisor(
-            phase=WorkflowPhase.CONTEXT_RETRIEVAL,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                CONTEXT_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="context_retriever",
-                    agent_role="context_retriever",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="context",
-                    node_name="finalize",
-                ),
-            },
-            self._context.build_state_update(result),
-            decision,
-        )
-        merged.pop(CONTEXT_AGENT_LOCAL_KEY, None)
-        merged.pop(CONTEXT_SELECTION_OUTPUT_KEY, None)
-        merged.pop(CONTEXT_SUFFICIENCY_OUTPUT_KEY, None)
-        merged.pop("evidence_drafts", None)
-        merged.pop("llm_provider_result", None)
-        return merged
-
-    def _analysis_subgraph_init_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_planning")
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="work_analysis",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={
-                "request_intent": _require_state_value(state["request_intent"], "request_intent"),
-                "context_result": _require_state_value(state["context_result"], "context_result"),
-            },
-            prompt_ref=self._analysis.analyze_prompt_ref,
-        )
-        return {
-            **state,
-            ANALYSIS_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="work_analysis",
-                agent_role="work_analysis",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="analysis",
-                node_name="init",
-                prompt_ref=self._analysis.analyze_prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _analysis_subgraph_analyze_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[ANALYSIS_AGENT_LOCAL_KEY])
-        llm_result = self._analysis.invoke_analyze_llm(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            context_result=_require_state_value(state["context_result"], "context_result"),
-            request=request,
-        )
-        result = self._analysis.build_output_from_llm_result(
-            llm_result,
-            context_result=_require_state_value(state["context_result"], "context_result"),
-        )
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "ANALYZE_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            ANALYSIS_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "analysis_result": result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="work_analysis",
-                agent_role="work_analysis",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="analysis",
-                node_name="analyze",
-                llm_call_id=f"{request.run_id}:analysis.analyze",
-                prompt_ref=self._analysis.analyze_prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _analysis_subgraph_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[ANALYSIS_AGENT_LOCAL_KEY])
-        result = _require_state_value(state["analysis_result"], "analysis_result")
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            ANALYSIS_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "analysis_result": result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="work_analysis",
-                agent_role="work_analysis",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="analysis",
-                node_name="result_validate",
-            ),
-        }
-
-    def _analysis_subgraph_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[ANALYSIS_AGENT_LOCAL_KEY])
-        result = _require_state_value(state["analysis_result"], "analysis_result")
-        decision = route_supervisor(
-            phase=WorkflowPhase.WORK_ANALYSIS,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                ANALYSIS_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="work_analysis",
-                    agent_role="work_analysis",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="analysis",
-                    node_name="finalize",
-                ),
-            },
-            self._analysis.build_state_update(result),
-            decision,
-        )
-        merged.pop(ANALYSIS_AGENT_LOCAL_KEY, None)
-        return merged
-
-    def _planning_subgraph_init_node(self, state: GraphState) -> GraphState:
-        invocation_id = self._id_factory()
-        review = state["plan_review"]
-        request_intent = _require_state_value(state["request_intent"], "request_intent")
-        analysis_result = _require_state_value(state["analysis_result"], "analysis_result")
-        mode = self._planning_mode_from_request_intent(request_intent)
-        if review is not None and review.get("status") == ReviewResult.REVISE.value:
-            mode = "revise_answer" if state.get("answer_draft") is not None else "revise_plan"
-        prompt_ref = {
-            "answer_only": self._planning.answer_only_prompt_ref,
-            "draft_plan": self._planning.draft_plan_prompt_ref,
-            "revise_answer": self._planning.revise_answer_prompt_ref,
-            "revise_plan": self._planning.revise_plan_prompt_ref,
-        }[mode]
-        local_state = build_agent_local_state(
-            agent_role="planning",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={
-                "request_intent": request_intent,
-                "analysis_result": analysis_result,
-                "mode": mode,
-            },
-            prompt_ref=prompt_ref,
-        )
-        return {
-            **state,
-            PLANNING_AGENT_LOCAL_KEY: local_state,
-            PLANNING_MODE_KEY: mode,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="planning",
-                agent_role="planning",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="planning",
-                node_name="init",
-                prompt_ref=prompt_ref,
-                agent_invocation_increment=1,
-                revision_increment=1 if mode.startswith("revise") else 0,
-            ),
-        }
-
-    def _planning_subgraph_plan_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[PLANNING_AGENT_LOCAL_KEY])
-        mode = state[PLANNING_MODE_KEY]
-        review_state = state["plan_review"]
-        review_issues: list[dict[str, object]] = []
-        review_summary: str | None = None
-        if review_state is not None:
-            review_issues = [dict(issue) for issue in review_state["issues"]]
-            review_summary = review_state.get("summary")
-        result: AnswerDraftV1 | ActionPlanDraftV1
-        if mode == "answer_only":
-            llm_result = self._planning.invoke_answer_only_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                request=request,
-            )
-            result = self._planning.build_answer_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-            )
-            llm_call_id = f"{request.run_id}:planning.answer_only"
-        elif mode == "draft_plan":
-            llm_result = self._planning.invoke_draft_plan_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                request=request,
-            )
-            result = self._planning.build_plan_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-            )
-            llm_call_id = f"{request.run_id}:planning.draft_plan"
-        elif mode == "revise_answer":
-            llm_result = self._planning.invoke_revise_answer_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                answer_draft=_require_state_value(state["answer_draft"], "answer_draft"),
-                review_issues=review_issues,
-                review_summary=review_summary,
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                request=request,
-            )
-            result = self._planning.build_answer_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-            )
-            llm_call_id = f"{request.run_id}:planning.revise_answer"
-        else:
-            llm_result = self._planning.invoke_revise_plan_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                plan_draft=_require_state_value(state["plan_draft"], "plan_draft"),
-                review_issues=review_issues,
-                review_summary=review_summary,
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                request=request,
-            )
-            result = self._planning.build_plan_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-            )
-            llm_call_id = f"{request.run_id}:planning.revise_plan"
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "PLAN_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            PLANNING_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "__planning_result__": result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="planning",
-                agent_role="planning",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="planning",
-                node_name="plan",
-                llm_call_id=llm_call_id,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _planning_subgraph_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PLANNING_AGENT_LOCAL_KEY])
-        result = state["__planning_result__"]
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = result
-        return {
-            **state,
-            PLANNING_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="planning",
-                agent_role="planning",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="planning",
-                node_name="result_validate",
-            ),
-        }
-
-    def _planning_subgraph_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PLANNING_AGENT_LOCAL_KEY])
-        mode = state[PLANNING_MODE_KEY]
-        result = state["__planning_result__"]
-        if "answer" in result:
-            answer_result = validate_answer_draft_v1(
-                result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-            )
-            state_update = self._planning.build_answer_state_update(answer_result)
-        else:
-            plan_result = validate_action_plan_draft_v1(
-                result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-            )
-            state_update = self._planning.build_plan_state_update(plan_result)
-        decision = route_supervisor(
-            phase=WorkflowPhase.SOLUTION_PLANNING,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                PLANNING_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="planning",
-                    agent_role="planning",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="planning",
-                    node_name="finalize",
-                    revision_increment=1 if mode == "revise_plan" else 0,
-                ),
-            },
-            state_update,
-            decision,
-        )
-        merged.pop(PLANNING_AGENT_LOCAL_KEY, None)
-        merged.pop(PLANNING_MODE_KEY, None)
-        merged.pop("__planning_result__", None)
-        return merged
-
-    def _review_subgraph_init_node(self, state: GraphState) -> GraphState:
-        invocation_id = self._id_factory()
-        review = state["plan_review"]
-        mode = (
-            "recheck"
-            if review is not None and review.get("status") == ReviewResult.REVISE.value
-            else "inspect"
-        )
-        prompt_ref = (
-            self._review.recheck_prompt_ref
-            if mode == "recheck"
-            else self._review.inspect_prompt_ref
-        )
-        local_state = build_agent_local_state(
-            agent_role="review",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={
-                "mode": mode,
-                "has_answer_draft": state.get("answer_draft") is not None,
-                "has_plan_draft": state.get("plan_draft") is not None,
-            },
-            prompt_ref=prompt_ref,
-        )
-        return {
-            **state,
-            REVIEW_AGENT_LOCAL_KEY: local_state,
-            REVIEW_MODE_KEY: mode,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="review",
-                agent_role="review",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="review",
-                node_name="init",
-                prompt_ref=prompt_ref,
-                agent_invocation_increment=1,
-                revision_increment=1 if mode == "recheck" else 0,
-            ),
-        }
-
-    def _review_subgraph_review_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[REVIEW_AGENT_LOCAL_KEY])
-        mode = state[REVIEW_MODE_KEY]
-        if mode == "recheck":
-            llm_result = self._review.invoke_recheck_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                request=request,
-                deterministic_action_risks=state.get("__modify_review_risks__"),
-            )
-            result = self._review.build_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                allowed_statuses=frozenset({ReviewResult.PASS.value, ReviewResult.BLOCK.value}),
-            )
-            llm_call_id = f"{request.run_id}:review.recheck"
-        else:
-            llm_result = self._review.invoke_inspect_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                request=request,
-                deterministic_action_risks=state.get("__modify_review_risks__"),
-            )
-            result = self._review.build_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-            )
-            llm_call_id = f"{request.run_id}:review.inspect"
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "REVIEW_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            REVIEW_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "plan_review": result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="review",
-                agent_role="review",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="review",
-                node_name="review",
-                llm_call_id=llm_call_id,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _review_subgraph_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[REVIEW_AGENT_LOCAL_KEY])
-        result = _require_state_value(state["plan_review"], "plan_review")
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            REVIEW_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "plan_review": result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="review",
-                agent_role="review",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="review",
-                node_name="result_validate",
-            ),
-        }
-
-    def _review_subgraph_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[REVIEW_AGENT_LOCAL_KEY])
-        result = _require_state_value(state["plan_review"], "plan_review")
-        decision = route_supervisor(
-            phase=WorkflowPhase.PLAN_REVIEW,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                REVIEW_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="review",
-                    agent_role="review",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="review",
-                    node_name="finalize",
-                ),
-            },
-            self._review.build_state_update(result),
-            decision,
-        )
-        merged.pop(REVIEW_AGENT_LOCAL_KEY, None)
-        merged.pop(REVIEW_MODE_KEY, None)
-        return merged
-
-    def _three_stage_one_init_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "start_analysis")
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="request_source_agent",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection=self._profile_request_source_prompt_input(request),
-            prompt_ref=self._three_stage1_prompt_ref,
-        )
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_one",
-                agent_role="request_source_agent",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="three.stage1",
-                node_name="init",
-                prompt_ref=self._three_stage1_prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _three_stage_one_request_source_node(self, state: GraphState) -> GraphState:
-        assert self._three_stage1_prompt_ref is not None
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        llm_result = self._request_understanding._llm_runtime.invoke_structured(
-            prompt_ref=self._three_stage1_prompt_ref,
-            prompt_input=self._profile_request_source_prompt_input(request),
-            output_schema=PROFILE_REQUEST_SOURCE_OUTPUT_SCHEMA,
-            trace_context=self._profile_trace_context(
-                request=request,
-                llm_call_id=f"{request.run_id}:profile.three.stage1.initial",
-            ),
-        )
-        output = validate_profile_request_source_output_v1(llm_result.structured_output)
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "REQUEST_SOURCE_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], output)
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            PROFILE_REQUEST_SOURCE_OUTPUT_KEY: output,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_one",
-                agent_role="request_source_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage1",
-                node_name="request_source",
-                llm_call_id=f"{request.run_id}:profile.three.stage1.initial",
-                prompt_ref=self._three_stage1_prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _three_stage_one_plan_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        prompt_output = state[PROFILE_REQUEST_SOURCE_OUTPUT_KEY]
-        source_plan = prompt_output["source_plan"]
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "PLAN_VALIDATED"
-        updated_local["typed_result"] = prompt_output
-        next_state: GraphState = {
-            **state,
-            "request_intent": prompt_output["request_intent"],
-            "source_fetch_plans": source_plan["source_fetch_plans"],
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_one",
-                agent_role="request_source_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage1",
-                node_name="plan_validate",
-            ),
-        }
-        if source_plan["result"] == "NO_FETCH_NEEDED":
-            next_state["acquisition_result"] = self._build_no_fetch_acquisition_result()
-        return next_state
-
-    def _route_profile_stage_one_plan_validate(self, state: GraphState) -> str:
-        prompt_output = state[PROFILE_REQUEST_SOURCE_OUTPUT_KEY]
-        source_plan = prompt_output["source_plan"]
-        return "deterministic_read" if source_plan["result"] == "PLAN_READY" else "finalize"
-
-    def _route_single_workflow_plan_validate(self, state: GraphState) -> str:
-        prompt_output = state[PROFILE_REQUEST_SOURCE_OUTPUT_KEY]
-        source_plan = prompt_output["source_plan"]
-        if source_plan["result"] == "PLAN_READY":
-            return "deterministic_read"
-        if source_plan["result"] == "NO_FETCH_NEEDED":
-            return "reason_plan"
-        return "finalize"
-
-    def _three_stage_one_read_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_retrieval")
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        result = self._acquisition.acquire(
-            plans=state["source_fetch_plans"],
-            request=request,
-            request_intent=state.get("request_intent"),
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "READ_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            "acquisition_result": result,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_one",
-                agent_role="request_source_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage1",
-                node_name="deterministic_read",
-            ),
-        }
-
-    def _three_stage_one_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        acquisition_result = validate_acquisition_result_v1(state["acquisition_result"])
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = cast(dict[str, object], acquisition_result)
-        return {
-            **state,
-            "acquisition_result": acquisition_result,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_one",
-                agent_role="request_source_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage1",
-                node_name="result_validate",
-            ),
-        }
-
-    def _three_stage_one_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        prompt_output = state[PROFILE_REQUEST_SOURCE_OUTPUT_KEY]
-        source_plan = prompt_output["source_plan"]
-        request_intent = prompt_output["request_intent"]
-        current: GraphState = {
-            **state,
-            "request_intent": request_intent,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_one",
-                agent_role="request_source_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage1",
-                node_name="finalize",
-            ),
-        }
-        if source_plan["result"] != "PLAN_READY":
-            decision = route_supervisor(
-                phase=WorkflowPhase.SOURCE_PLANNING,
-                state=cast(MultiAgentGraphState, current),
-                result=source_plan,
-            )
-            updated_local = dict(local_state)
-            updated_local["node_state"] = "FINALIZED"
-            updated_local["disposition"] = {
-                "schema_version": 1,
-                "status": cast(str, source_plan["result"]),
-                "next_target": cast(str, decision["target"]),
-                "reason_code": cast(str | None, decision.get("reason_code")),
-            }
-            merged = self._merge_decision(
-                {**current, PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local)},
-                {
-                    "request_intent": request_intent,
-                    **self._acquisition.build_planning_state_update(source_plan),
-                },
-                decision,
-            )
-        else:
-            acquisition_result = _require_state_value(
-                state["acquisition_result"], "acquisition_result"
-            )
-            decision = route_supervisor(
-                phase=WorkflowPhase.API_ACQUISITION,
-                state=cast(MultiAgentGraphState, current),
-                result=acquisition_result,
-            )
-            updated_local = dict(local_state)
-            updated_local["node_state"] = "FINALIZED"
-            updated_local["disposition"] = {
-                "schema_version": 1,
-                "status": cast(str, acquisition_result["status"]),
-                "next_target": cast(str, decision["target"]),
-                "reason_code": cast(str | None, decision.get("reason_code")),
-            }
-            merged = self._merge_decision(
-                {**current, PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local)},
-                {
-                    "request_intent": request_intent,
-                    **self._acquisition.build_planning_state_update(source_plan),
-                    **self._acquisition.build_acquisition_state_update(acquisition_result),
-                },
-                decision,
-            )
-        merged.pop(PROFILE_AGENT_LOCAL_KEY, None)
-        merged.pop(PROFILE_REQUEST_SOURCE_OUTPUT_KEY, None)
-        return merged
-
-    def _three_stage_two_init_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_planning")
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="evidence_analysis_plan_agent",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection=self._profile_post_read_prompt_input(state),
-            prompt_ref=self._three_stage2_prompt_ref,
-        )
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_two",
-                agent_role="evidence_analysis_plan_agent",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="three.stage2",
-                node_name="init",
-                prompt_ref=self._three_stage2_prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _three_stage_two_reason_plan_node(self, state: GraphState) -> GraphState:
-        assert self._three_stage2_prompt_ref is not None
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        llm_result = self._request_understanding._llm_runtime.invoke_structured(
-            prompt_ref=self._three_stage2_prompt_ref,
-            prompt_input=self._profile_post_read_prompt_input(state),
-            output_schema=PROFILE_FUSED_PLANNING_OUTPUT_SCHEMA,
-            trace_context=self._profile_trace_context(
-                request=request,
-                llm_call_id=f"{request.run_id}:profile.three.stage2.initial",
-            ),
-        )
-        output = validate_profile_reason_plan_output_v1(llm_result.structured_output)
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "REASON_PLAN_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], output)
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            PROFILE_REASON_PLAN_OUTPUT_KEY: output,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_two",
-                agent_role="evidence_analysis_plan_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage2",
-                node_name="reason_plan",
-                llm_call_id=f"{request.run_id}:profile.three.stage2.initial",
-                prompt_ref=self._three_stage2_prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _three_stage_two_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        output = state[PROFILE_REASON_PLAN_OUTPUT_KEY]
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = output
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_two",
-                agent_role="evidence_analysis_plan_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage2",
-                node_name="result_validate",
-            ),
-        }
-
-    def _three_stage_two_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        output = state[PROFILE_REASON_PLAN_OUTPUT_KEY]
-        context_result = output["context_result"]
-        analysis_result = output["analysis_result"]
-        planning_result = output["planning_result"]
-        result = self._planning_result_from_projection(planning_result)
-        state_update = self._profile_reason_plan_state_update(output)
-        decision = route_supervisor(
-            phase=WorkflowPhase.SOLUTION_PLANNING,
-            state=cast(
-                MultiAgentGraphState,
-                {
-                    **state,
-                    "context_result": context_result,
-                    "analysis_result": analysis_result,
-                },
-            ),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="stage_two",
-                    agent_role="evidence_analysis_plan_agent",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="three.stage2",
-                    node_name="finalize",
-                ),
-            },
-            state_update,
-            decision,
-        )
-        merged.pop(PROFILE_AGENT_LOCAL_KEY, None)
-        merged.pop(PROFILE_REASON_PLAN_OUTPUT_KEY, None)
-        return merged
-
-    def _three_stage_review_init_node(self, state: GraphState) -> GraphState:
-        invocation_id = self._id_factory()
-        review = state["plan_review"]
-        mode = (
-            "recheck"
-            if review is not None and review.get("status") == ReviewResult.REVISE.value
-            else "inspect"
-        )
-        prompt_ref = (
-            self._review.recheck_prompt_ref
-            if mode == "recheck"
-            else self._review.inspect_prompt_ref
-        )
-        local_state = build_agent_local_state(
-            agent_role="review",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection={"mode": mode},
-            prompt_ref=prompt_ref,
-        )
-        return {
-            **state,
-            REVIEW_AGENT_LOCAL_KEY: local_state,
-            REVIEW_MODE_KEY: mode,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_three",
-                agent_role="review",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="three.stage3",
-                node_name="init",
-                prompt_ref=prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _three_stage_review_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[REVIEW_AGENT_LOCAL_KEY])
-        mode = state[REVIEW_MODE_KEY]
-        if mode == "recheck":
-            llm_result = self._review.invoke_recheck_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                request=request,
-            )
-            result = self._review.build_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                allowed_statuses=frozenset({ReviewResult.PASS.value, ReviewResult.BLOCK.value}),
-            )
-            llm_call_id = f"{request.run_id}:review.recheck"
-        else:
-            llm_result = self._review.invoke_inspect_llm(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                request=request,
-            )
-            result = self._review.build_output_from_llm_result(
-                llm_result,
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-            )
-            llm_call_id = f"{request.run_id}:review.inspect"
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "REVIEW_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            REVIEW_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "plan_review": result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_three",
-                agent_role="review",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage3",
-                node_name="review",
-                llm_call_id=llm_call_id,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _three_stage_review_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[REVIEW_AGENT_LOCAL_KEY])
-        result = _require_state_value(state["plan_review"], "plan_review")
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = result
-        return {
-            **state,
-            REVIEW_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="stage_three",
-                agent_role="review",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="three.stage3",
-                node_name="result_validate",
-            ),
-        }
-
-    def _three_stage_review_finalize_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[REVIEW_AGENT_LOCAL_KEY])
-        result = _require_state_value(state["plan_review"], "plan_review")
-        decision = route_supervisor(
-            phase=WorkflowPhase.PLAN_REVIEW,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                REVIEW_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="stage_three",
-                    agent_role="review",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="three.stage3",
-                    node_name="finalize",
-                ),
-            },
-            self._review.build_state_update(result),
-            decision,
-        )
-        merged.pop(REVIEW_AGENT_LOCAL_KEY, None)
-        merged.pop(REVIEW_MODE_KEY, None)
-        return merged
-
-    def _single_workflow_init_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "start_analysis")
-        invocation_id = self._id_factory()
-        local_state = build_agent_local_state(
-            agent_role="unified_agent",
-            invocation_id=invocation_id,
-            node_state="INITIALIZED",
-            input_projection=self._profile_request_source_prompt_input(request),
-            prompt_ref=self._single_request_source_prompt_ref,
-        )
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: local_state,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=invocation_id,
-                subgraph_namespace="single",
-                node_name="init",
-                prompt_ref=self._single_request_source_prompt_ref,
-                agent_invocation_increment=1,
-            ),
-        }
-
-    def _single_workflow_request_source_node(self, state: GraphState) -> GraphState:
-        assert self._single_request_source_prompt_ref is not None
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        llm_result = self._request_understanding._llm_runtime.invoke_structured(
-            prompt_ref=self._single_request_source_prompt_ref,
-            prompt_input=self._profile_request_source_prompt_input(request),
-            output_schema=PROFILE_REQUEST_SOURCE_OUTPUT_SCHEMA,
-            trace_context=self._profile_trace_context(
-                request=request,
-                llm_call_id=f"{request.run_id}:profile.single.request_source.initial",
-            ),
-        )
-        output = validate_profile_request_source_output_v1(llm_result.structured_output)
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "REQUEST_SOURCE_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], output)
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            PROFILE_REQUEST_SOURCE_OUTPUT_KEY: output,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="single",
-                node_name="request_source",
-                llm_call_id=f"{request.run_id}:profile.single.request_source.initial",
-                prompt_ref=self._single_request_source_prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _single_workflow_plan_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        prompt_output = state[PROFILE_REQUEST_SOURCE_OUTPUT_KEY]
-        source_plan = prompt_output["source_plan"]
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "PLAN_VALIDATED"
-        updated_local["typed_result"] = prompt_output
-        next_state: GraphState = {
-            **state,
-            "request_intent": prompt_output["request_intent"],
-            "source_fetch_plans": source_plan["source_fetch_plans"],
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="single",
-                node_name="plan_validate",
-            ),
-        }
-        if source_plan["result"] == "NO_FETCH_NEEDED":
-            next_state["acquisition_result"] = self._build_no_fetch_acquisition_result()
-        return next_state
-
-    def _single_workflow_read_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_retrieval")
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        result = self._acquisition.acquire(
-            plans=state["source_fetch_plans"],
-            request=request,
-            request_intent=state.get("request_intent"),
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "READ_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], result)
-        return {
-            **state,
-            "acquisition_result": result,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="single",
-                node_name="deterministic_read",
-            ),
-        }
-
-    def _single_workflow_reason_plan_node(self, state: GraphState) -> GraphState:
-        assert self._single_reason_plan_prompt_ref is not None
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_planning")
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        llm_result = self._request_understanding._llm_runtime.invoke_structured(
-            prompt_ref=self._single_reason_plan_prompt_ref,
-            prompt_input=self._profile_post_read_prompt_input(state),
-            output_schema=PROFILE_FUSED_PLANNING_OUTPUT_SCHEMA,
-            trace_context=self._profile_trace_context(
-                request=request,
-                llm_call_id=f"{request.run_id}:profile.single.reason_plan.initial",
-            ),
-        )
-        output = validate_profile_reason_plan_output_v1(llm_result.structured_output)
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "REASON_PLAN_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], output)
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            PROFILE_REASON_PLAN_OUTPUT_KEY: output,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="single",
-                node_name="reason_plan",
-                llm_call_id=f"{request.run_id}:profile.single.reason_plan.initial",
-                prompt_ref=self._single_reason_plan_prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _single_workflow_self_review_node(self, state: GraphState) -> GraphState:
-        assert self._single_review is not None
-        request = self._request_from_state(state)
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        output = state[PROFILE_REASON_PLAN_OUTPUT_KEY]
-        planning_result = output["planning_result"]
-        result = self._planning_result_from_projection(planning_result)
-        answer_draft = (
-            validate_answer_draft_v1(result, analysis_result=output["analysis_result"])
-            if "answer" in result
-            else None
-        )
-        plan_draft = (
-            validate_action_plan_draft_v1(result, analysis_result=output["analysis_result"])
-            if "plan_id" in result
-            else None
-        )
-        llm_result = self._single_review.invoke_inspect_llm(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            context_result=output["context_result"],
-            analysis_result=output["analysis_result"],
-            answer_draft=answer_draft,
-            plan_draft=plan_draft,
-            request=request,
-            deterministic_action_risks=state.get("__modify_review_risks__"),
-        )
-        review_result = self._single_review.build_output_from_llm_result(
-            llm_result,
-            analysis_result=output["analysis_result"],
-            answer_draft=answer_draft,
-            plan_draft=plan_draft,
-        )
-        updated_local = dict(record_llm_result(local_state, llm_result))
-        updated_local["node_state"] = "SELF_REVIEW_COMPLETE"
-        updated_local["typed_result"] = cast(dict[str, object], review_result)
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "context_result": output["context_result"],
-            "analysis_result": output["analysis_result"],
-            **self._profile_planning_state_update(
-                planning_result,
-                analysis_result=output["analysis_result"],
-            ),
-            "plan_review": review_result,
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="single",
-                node_name="self_review",
-                llm_call_id=f"{request.run_id}:profile.single.self_review.initial",
-                prompt_ref=self._single_review.inspect_prompt_ref,
-                llm_call_increment=llm_result.structured_output_attempts,
-                repair_increment=max(0, llm_result.structured_output_attempts - 1),
-            ),
-        }
-
-    def _single_workflow_result_validate_node(self, state: GraphState) -> GraphState:
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        review_result = _require_state_value(state["plan_review"], "plan_review")
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "RESULT_VALIDATED"
-        updated_local["typed_result"] = review_result
-        return {
-            **state,
-            PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-            "trace_context": merge_trace_context(
-                state,
-                graph_profile=self._graph_profile.value,
-                agent_subgraph_id="single_workflow",
-                agent_role="unified_agent",
-                agent_invocation_id=local_state["invocation_id"],
-                subgraph_namespace="single",
-                node_name="result_validate",
-            ),
-        }
-
-    def _single_workflow_finalize_node(self, state: GraphState) -> GraphState:
-        assert self._single_review is not None
-        local_state = cast(AgentLocalStateV1, state[PROFILE_AGENT_LOCAL_KEY])
-        if state.get("plan_review") is None:
-            prompt_output = state[PROFILE_REQUEST_SOURCE_OUTPUT_KEY]
-            source_plan = prompt_output["source_plan"]
-            request_intent = prompt_output["request_intent"]
-            current: GraphState = {
-                **state,
-                "request_intent": request_intent,
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="single_workflow",
-                    agent_role="unified_agent",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="single",
-                    node_name="finalize",
-                ),
-            }
-            decision = route_supervisor(
-                phase=WorkflowPhase.SOURCE_PLANNING,
-                state=cast(MultiAgentGraphState, current),
-                result=source_plan,
-            )
-            updated_local = dict(local_state)
-            updated_local["node_state"] = "FINALIZED"
-            updated_local["disposition"] = {
-                "schema_version": 1,
-                "status": cast(str, source_plan["result"]),
-                "next_target": cast(str, decision["target"]),
-                "reason_code": cast(str | None, decision.get("reason_code")),
-            }
-            merged = self._merge_decision(
-                {**current, PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local)},
-                {
-                    "request_intent": request_intent,
-                    **self._acquisition.build_planning_state_update(source_plan),
-                },
-                decision,
-            )
-            merged.pop(PROFILE_AGENT_LOCAL_KEY, None)
-            merged.pop(PROFILE_REQUEST_SOURCE_OUTPUT_KEY, None)
-            return merged
-        result = _require_state_value(state["plan_review"], "plan_review")
-        decision = route_supervisor(
-            phase=WorkflowPhase.PLAN_REVIEW,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        updated_local = dict(local_state)
-        updated_local["node_state"] = "FINALIZED"
-        updated_local["disposition"] = {
-            "schema_version": 1,
-            "status": cast(str, result["status"]),
-            "next_target": cast(str, decision["target"]),
-            "reason_code": cast(str | None, decision.get("reason_code")),
-        }
-        merged = self._merge_decision(
-            {
-                **state,
-                PROFILE_AGENT_LOCAL_KEY: cast(AgentLocalStateV1, updated_local),
-                "trace_context": merge_trace_context(
-                    state,
-                    graph_profile=self._graph_profile.value,
-                    agent_subgraph_id="single_workflow",
-                    agent_role="unified_agent",
-                    agent_invocation_id=local_state["invocation_id"],
-                    subgraph_namespace="single",
-                    node_name="finalize",
-                ),
-            },
-            self._single_review.build_state_update(result),
-            decision,
-        )
-        merged.pop(PROFILE_AGENT_LOCAL_KEY, None)
-        merged.pop(PROFILE_REASON_PLAN_OUTPUT_KEY, None)
-        return merged
-
-    def _profile_request_source_prompt_input(
-        self,
-        request: WorkflowStartRequest,
-    ) -> dict[str, object]:
-        return {
-            "request_text": request.request_text,
-            "entry_mode": request.entry_mode,
-            "selected_resource_ids": list(request.selected_resource_ids),
-        }
-
-    def _profile_post_read_prompt_input(self, state: GraphState) -> dict[str, object]:
-        request = self._request_from_state(state)
-        return {
-            "request_text": request.request_text,
-            "request_intent": _require_state_value(state["request_intent"], "request_intent"),
-            "acquisition_result": _require_state_value(
-                state["acquisition_result"], "acquisition_result"
-            ),
-        }
-
-    def _profile_trace_context(
-        self,
-        *,
-        request: WorkflowStartRequest,
-        llm_call_id: str,
-    ) -> ObservabilityContext:
-        return ObservabilityContext(
-            request_id=request.correlation.request_id,
-            command_id=request.correlation.command_id,
-            conversation_id=request.conversation_id,
-            run_id=request.run_id,
-            langgraph_thread_id=request.workflow_key,
-            llm_call_id=llm_call_id,
-        )
-
-    def _planning_result_from_projection(
-        self,
-        planning_result: ProfilePlanningProjectionV1,
-    ) -> AnswerDraftV1 | ActionPlanDraftV1:
-        answer_draft = planning_result["answer_draft"]
-        if answer_draft is not None:
-            return answer_draft
-        plan_draft = planning_result["plan_draft"]
-        if plan_draft is not None:
-            return plan_draft
-        raise ValueError("planning_result must contain answer_draft or plan_draft")
-
-    def _profile_reason_plan_state_update(
-        self,
-        output: ProfileReasonPlanOutputV1,
-    ) -> GraphStateUpdateV1:
-        planning_result = output["planning_result"]
-        return {
-            "context_result": output["context_result"],
-            "analysis_result": output["analysis_result"],
-            **self._profile_planning_state_update(
-                planning_result,
-                analysis_result=output["analysis_result"],
-            ),
-        }
-
-    def _profile_planning_state_update(
-        self,
-        planning_result: ProfilePlanningProjectionV1,
-        *,
-        analysis_result: WorkAnalysisResultV1,
-    ) -> GraphStateUpdateV1:
-        result = self._planning_result_from_projection(planning_result)
-        if "answer" in result:
-            answer_result = validate_answer_draft_v1(result, analysis_result=analysis_result)
-            return self._planning.build_answer_state_update(answer_result)
-        plan_result = validate_action_plan_draft_v1(result, analysis_result=analysis_result)
-        return self._planning.build_plan_state_update(plan_result)
-
-    def _build_no_fetch_acquisition_result(self) -> AcquisitionResultV1:
-        return {
-            "schema_version": 1,
-            "status": "COMPLETE",
-            "resource_handles": [],
-            "source_summaries": [],
-            "missing_slots": [],
-            "remaining_budget": {
-                "sources": 0,
-                "pages": 0,
-                "candidates": 0,
-                "details": 0,
-            },
-        }
-
-    def _request_understanding_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "start_analysis")
-        output = self._request_understanding.classify(request)
-        decision = route_supervisor(
-            phase=WorkflowPhase.REQUEST_ANALYSIS,
-            state=cast(MultiAgentGraphState, state),
-            result=output,
-        )
-        return self._merge_decision(
-            state,
-            self._request_understanding.build_state_update(output, request=request),
-            decision,
-        )
-
-    def _source_planning_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_retrieval")
-        additional: AdditionalAcquisitionRequestV1 | None = None
-        context_result = state.get("context_result")
-        analysis_result = state.get("analysis_result")
-        if context_result is not None:
-            additional = context_result["additional_acquisition_request"]
-        if additional is None and analysis_result is not None:
-            additional = analysis_result["additional_acquisition_request"]
-        output = self._acquisition.plan_sources(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            request=request,
-            additional_acquisition_request=additional,
-        )
-        decision = route_supervisor(
-            phase=WorkflowPhase.SOURCE_PLANNING,
-            state=cast(MultiAgentGraphState, state),
-            result=output,
-        )
-        return self._merge_decision(
-            state, self._acquisition.build_planning_state_update(output), decision
-        )
-
-    def _api_acquisition_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        result = self._acquisition.acquire(
-            plans=state["source_fetch_plans"],
-            request=request,
-            request_intent=state.get("request_intent"),
-        )
-        decision = route_supervisor(
-            phase=WorkflowPhase.API_ACQUISITION,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        return self._merge_decision(
-            state, self._acquisition.build_acquisition_state_update(result), decision
-        )
-
-    def _context_retrieval_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        result = self._context.retrieve(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            acquisition_result=_require_state_value(
-                state["acquisition_result"], "acquisition_result"
-            ),
-            request=request,
-        )
-        decision = route_supervisor(
-            phase=WorkflowPhase.CONTEXT_RETRIEVAL,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        return self._merge_decision(state, self._context.build_state_update(result), decision)
-
-    def _work_analysis_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        self._transition_run(request.run_id, "begin_planning")
-        result = self._analysis.analyze(
-            request_intent=_require_state_value(state["request_intent"], "request_intent"),
-            context_result=_require_state_value(state["context_result"], "context_result"),
-            request=request,
-        )
-        decision = route_supervisor(
-            phase=WorkflowPhase.WORK_ANALYSIS,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        return self._merge_decision(state, self._analysis.build_state_update(result), decision)
-
-    def _plan_review_node(self, state: GraphState) -> GraphState:
-        request = self._request_from_state(state)
-        review = state["plan_review"]
-        if review is not None and review.get("status") == ReviewResult.REVISE.value:
-            result = self._review.recheck(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                request=request,
-            )
-        else:
-            result = self._review.inspect(
-                request_intent=_require_state_value(state["request_intent"], "request_intent"),
-                context_result=_require_state_value(state["context_result"], "context_result"),
-                analysis_result=_require_state_value(state["analysis_result"], "analysis_result"),
-                answer_draft=state["answer_draft"],
-                plan_draft=state["plan_draft"],
-                request=request,
-            )
-        decision = route_supervisor(
-            phase=WorkflowPhase.PLAN_REVIEW,
-            state=cast(MultiAgentGraphState, state),
-            result=result,
-        )
-        return self._merge_decision(state, self._review.build_state_update(result), decision)
 
     def _domain_validation_node(self, state: GraphState) -> GraphState:
         plan_draft = _require_state_value(state["plan_draft"], "plan_draft")
@@ -3822,38 +1584,7 @@ class LangGraphWorkflowRuntime(WorkflowRuntime):
         return "source_planning"
 
     def _request_from_state(self, state: GraphState) -> WorkflowStartRequest:
-        request = state.get("__request__")
-        if not isinstance(request, WorkflowStartRequest):
-            raise TypeError("workflow state is missing WorkflowStartRequest")
-        prompt_context = cast(dict[str, object], state.get("prompt_context", {}))
-        confirmation_response = prompt_context.get("confirmation_response")
-        if not isinstance(confirmation_response, dict):
-            return request
-        request_text = (
-            request.request_text
-            + "\n\n[clarification]\n"
-            + dumps(
-                {
-                    "selected_option_ids": cast(
-                        list[str], confirmation_response.get("selected_option_ids", [])
-                    ),
-                    "free_text": cast(str | None, confirmation_response.get("free_text")),
-                },
-                ensure_ascii=True,
-                sort_keys=True,
-            )
-        )
-        return WorkflowStartRequest(
-            run_id=request.run_id,
-            conversation_id=request.conversation_id,
-            workflow_key=request.workflow_key,
-            entry_mode=request.entry_mode,
-            requested_mode=request.requested_mode,
-            request_text=request_text,
-            selected_resource_ids=request.selected_resource_ids,
-            correlation=request.correlation,
-            selected_resources=request.selected_resources,
-        )
+        return request_from_state(state)
 
     def _config_for_thread(self, workflow_key: str) -> dict[str, object]:
         return {"configurable": {"thread_id": workflow_key}}
@@ -4412,26 +2143,4 @@ class LangGraphWorkflowRuntime(WorkflowRuntime):
         return sha256(dumps(payload, sort_keys=True).encode("utf-8")).hexdigest()
 
     def _planning_mode_from_request_intent(self, request_intent: RequestIntentV1) -> str:
-        """Deterministic answer_only/draft_plan selection (GAP-F1).
-
-        Reads the ``response_disposition`` typed field produced by
-        ``request_understanding.classify`` instead of matching keyword
-        substrings against ``request_text`` -- the previous approach silently
-        misrouted any request whose natural-language phrasing (in particular
-        Korean) did not contain one of a fixed English token list. The field
-        is optional on ``RequestIntentV1`` (see its definition) because only
-        SIX_ROLE_BASELINE's standalone Planning subgraph needs this upfront
-        choice between the two mutually exclusive ``planning.answer_only`` /
-        ``planning.draft_plan`` prompt slots; SINGLE_BASELINE and THREE_STAGE
-        decide ANSWER_ONLY vs PLAN_READY inside one fused planning call. When
-        the field is absent -- an older classify prompt version that predates
-        this field, or a profile that never sets it -- this falls back to
-        ``answer_only`` rather than guessing an action the user did not ask
-        for (docs/01-b-policy-definition-v2.8.md POL-EVD-003 / "Answer-only에서
-        불필요한 Action을 생성하지 않는다").
-        """
-        return (
-            "draft_plan"
-            if request_intent.get("response_disposition") == "ACTION_REQUIRED"
-            else "answer_only"
-        )
+        return planning_mode_from_request_intent(request_intent)
