+++ src/google_work_agent/application/workflows/api_acquisition.py
@@ -6,12 +6,14 @@ import time as _time_module
 from collections.abc import Callable
 from dataclasses import asdict, dataclass
 from datetime import UTC, datetime, time, timedelta
+from functools import partial
 from pathlib import Path
 from typing import Literal, Required, TypedDict, cast
 from zoneinfo import ZoneInfo, ZoneInfoNotFoundError
 
 from google_work_agent.application.llm import StructuredLLMRuntime
 from google_work_agent.application.observability import ObservabilityContext
+from google_work_agent.application.workflows import _schema_support as _schema
 from google_work_agent.application.workflows.contracts import (
     AdditionalAcquisitionRequestV1,
     ApiAcquisitionResult,
@@ -1466,23 +1468,6 @@ def _resource_handle(snapshot: ResourceSnapshot) -> str:
     return f"{snapshot.resource_type.value}:{snapshot.resource_id}"
 
 
-def _provider_summary(result: StructuredLLMResult) -> dict[str, object]:
-    return {
-        "provider": result.provider,
-        "model": result.model,
-        "requested_mode": result.requested_mode.value,
-        "actual_runtime": result.actual_runtime.value,
-        "input_tokens": result.input_tokens,
-        "output_tokens": result.output_tokens,
-        "total_tokens": result.total_tokens,
-        "latency_ms": result.latency_ms,
-        "fallback_reason": result.fallback_reason,
-        "structured_output_attempts": result.structured_output_attempts,
-        "provider_request_id": result.provider_request_id,
-        "safe_error_code": result.safe_error_code,
-    }
-
-
 def _validate_source_summaries(value: object) -> list[dict[str, object]]:
     summaries = _require_list(value, "$.source_summaries")
     validated: list[dict[str, object]] = []
@@ -1560,32 +1545,20 @@ def _validate_remaining_budget(value: object) -> dict[str, int]:
     }
 
 
-def _require_mapping(value: object, path: str) -> dict[str, object]:
-    if not isinstance(value, dict):
-        raise SourcePlanningValidationError(f"{path} must be an object")
-    result: dict[str, object] = {}
-    for key, item in value.items():
-        if not isinstance(key, str):
-            raise SourcePlanningValidationError(f"{path} keys must be strings")
-        result[key] = item
-    return result
-
-
-def _require_exact_keys(value: dict[str, object], path: str, keys: set[str]) -> None:
-    actual = set(value)
-    missing = keys - actual
-    extra = actual - keys
-    if missing:
-        raise SourcePlanningValidationError(f"{path} is missing required fields: {sorted(missing)}")
-    if extra:
-        raise SourcePlanningValidationError(f"{path} has unsupported fields: {sorted(extra)}")
-
-
-def _require_int(value: dict[str, object], field: str, path: str) -> int:
-    item = value[field]
-    if not isinstance(item, int) or isinstance(item, bool):
-        raise SourcePlanningValidationError(f"{path}.{field} must be integer")
-    return item
+# Shared with the other agent workflow modules; see _schema_support module docstring.
+_require_mapping = partial(_schema.require_mapping, error_cls=SourcePlanningValidationError)
+_require_exact_keys = partial(_schema.require_exact_keys, error_cls=SourcePlanningValidationError)
+_require_int = partial(_schema.require_int, error_cls=SourcePlanningValidationError)
+_require_string = partial(_schema.require_string, error_cls=SourcePlanningValidationError)
+_require_list = partial(_schema.require_list, error_cls=SourcePlanningValidationError)
+_require_string_list = partial(_schema.require_string_list, error_cls=SourcePlanningValidationError)
+_optional_string_list = partial(
+    _schema.optional_string_list, error_cls=SourcePlanningValidationError
+)
+_optional_option_list = partial(
+    _schema.optional_option_list, error_cls=SourcePlanningValidationError
+)
+_provider_summary = _schema.provider_summary
 
 
 def _require_positive_int(
@@ -1599,46 +1572,3 @@ def _require_positive_int(
     if item < minimum:
         raise SourcePlanningValidationError(f"{path}.{field} must be >= {minimum}")
     return item
-
-
-def _require_string(value: dict[str, object], field: str, path: str) -> str:
-    item = value[field]
-    if not isinstance(item, str):
-        raise SourcePlanningValidationError(f"{path}.{field} must be string")
-    return item
-
-
-def _require_list(value: object, path: str) -> list[object]:
-    if not isinstance(value, list):
-        raise SourcePlanningValidationError(f"{path} must be an array")
-    return value
-
-
-def _require_string_list(value: object, path: str) -> list[str]:
-    if not isinstance(value, list):
-        raise SourcePlanningValidationError(f"{path} must be an array")
-    for index, item in enumerate(value):
-        if not isinstance(item, str):
-            raise SourcePlanningValidationError(f"{path}[{index}] must be string")
-    return cast(list[str], value)
-
-
-def _optional_string_list(value: object) -> list[str]:
-    if value is None:
-        return []
-    if not isinstance(value, list):
-        raise SourcePlanningValidationError("clarification list field must be an array")
-    result: list[str] = []
-    for index, item in enumerate(value):
-        if not isinstance(item, str):
-            raise SourcePlanningValidationError(f"clarification list entry must be string: {index}")
-        result.append(item)
-    return result
-
-
-def _optional_option_list(value: object) -> list[dict[str, object]]:
-    if value is None:
-        return []
-    if not isinstance(value, list):
-        raise SourcePlanningValidationError("clarification options must be an array")
-    return [_require_mapping(item, "$.clarification.options[]") for item in value]
