+++ src/google_work_agent/application/workflows/request_understanding.py
@@ -2,11 +2,13 @@
 
 from __future__ import annotations
 
+from functools import partial
 from pathlib import Path
 from typing import Final, Literal, NotRequired, Required, TypedDict, cast
 
 from google_work_agent.application.llm import StructuredLLMRuntime
 from google_work_agent.application.observability import ObservabilityContext
+from google_work_agent.application.workflows import _schema_support as _schema
 from google_work_agent.application.workflows.contracts import (
     CONFIRMATION_ORIGIN_TARGETS,
     ConfirmationResponseV1,
@@ -771,23 +773,6 @@ def resolve_confirmation_origin_target(
     return question["origin_target"]
 
 
-def _provider_summary(result: StructuredLLMResult) -> dict[str, object]:
-    return {
-        "provider": result.provider,
-        "model": result.model,
-        "requested_mode": result.requested_mode.value,
-        "actual_runtime": result.actual_runtime.value,
-        "input_tokens": result.input_tokens,
-        "output_tokens": result.output_tokens,
-        "total_tokens": result.total_tokens,
-        "latency_ms": result.latency_ms,
-        "fallback_reason": result.fallback_reason,
-        "structured_output_attempts": result.structured_output_attempts,
-        "provider_request_id": result.provider_request_id,
-        "safe_error_code": result.safe_error_code,
-    }
-
-
 def _validate_goal(value: object) -> RequestIntentGoalV1:
     goal = _require_mapping(value, "$.goal")
     _require_exact_keys(goal, "$.goal", {"summary", "user_visible_objective"})
@@ -961,73 +946,21 @@ def _validate_unsupported_scope(value: object) -> RequestIntentUnsupportedScopeV
     }
 
 
-def _require_mapping(value: object, path: str) -> dict[str, object]:
-    if not isinstance(value, dict):
-        raise RequestUnderstandingValidationError(f"{path} must be an object")
-    result: dict[str, object] = {}
-    for key, item in value.items():
-        if not isinstance(key, str):
-            raise RequestUnderstandingValidationError(f"{path} keys must be strings")
-        result[key] = item
-    return result
-
-
-def _require_exact_keys(value: dict[str, object], path: str, keys: set[str]) -> None:
-    actual = set(value)
-    missing = keys - actual
-    extra = actual - keys
-    if missing:
-        raise RequestUnderstandingValidationError(
-            f"{path} is missing required fields: {sorted(missing)}"
-        )
-    if extra:
-        raise RequestUnderstandingValidationError(f"{path} has unsupported fields: {sorted(extra)}")
-
-
-def _require_allowed_keys(
-    value: dict[str, object],
-    path: str,
-    *,
-    required: set[str],
-    optional: set[str],
-) -> None:
-    actual = set(value)
-    missing = required - actual
-    extra = actual - required - optional
-    if missing:
-        raise RequestUnderstandingValidationError(
-            f"{path} is missing required fields: {sorted(missing)}"
-        )
-    if extra:
-        raise RequestUnderstandingValidationError(f"{path} has unsupported fields: {sorted(extra)}")
-
-
-def _require_int(value: dict[str, object], field: str, path: str) -> int:
-    item = value[field]
-    if not isinstance(item, int) or isinstance(item, bool):
-        raise RequestUnderstandingValidationError(f"{path}.{field} must be integer")
-    return item
-
-
-def _require_string(value: dict[str, object], field: str, path: str) -> str:
-    item = value[field]
-    if not isinstance(item, str):
-        raise RequestUnderstandingValidationError(f"{path}.{field} must be string")
-    return item
-
-
-def _require_list(value: object, path: str) -> list[object]:
-    if not isinstance(value, list):
-        raise RequestUnderstandingValidationError(f"{path} must be an array")
-    return value
-
-
-def _require_string_list(value: object, path: str) -> list[str]:
-    items = _require_list(value, path)
-    for index, item in enumerate(items):
-        if not isinstance(item, str):
-            raise RequestUnderstandingValidationError(f"{path}[{index}] must be string")
-    return cast(list[str], items)
+# Shared with the other agent workflow modules; see _schema_support module docstring.
+_require_mapping = partial(_schema.require_mapping, error_cls=RequestUnderstandingValidationError)
+_require_exact_keys = partial(
+    _schema.require_exact_keys, error_cls=RequestUnderstandingValidationError
+)
+_require_allowed_keys = partial(
+    _schema.require_allowed_keys, error_cls=RequestUnderstandingValidationError
+)
+_require_int = partial(_schema.require_int, error_cls=RequestUnderstandingValidationError)
+_require_string = partial(_schema.require_string, error_cls=RequestUnderstandingValidationError)
+_require_list = partial(_schema.require_list, error_cls=RequestUnderstandingValidationError)
+_require_string_list = partial(
+    _schema.require_string_list, error_cls=RequestUnderstandingValidationError
+)
+_provider_summary = _schema.provider_summary
 
 
 def _non_empty(value: str) -> bool:
