+++ src/google_work_agent/application/workflows/solution_planning.py
@@ -3,11 +3,13 @@
 from __future__ import annotations
 
 import copy
+from functools import partial
 from pathlib import Path
 from typing import Final, Literal, NotRequired, Required, TypedDict, cast
 
 from google_work_agent.application.llm import StructuredLLMRuntime
 from google_work_agent.application.observability import ObservabilityContext
+from google_work_agent.application.workflows import _schema_support as _schema
 from google_work_agent.application.workflows.context_retrieval import ContextRetrievalResultV1
 from google_work_agent.application.workflows.contracts import (
     GraphStateUpdateV1,
@@ -812,9 +814,7 @@ def _validate_action_draft(
     tool_name = _require_string(action, "tool_name", path)
     entry = registry.get(tool_name)
     if entry is None:
-        raise SolutionPlanningValidationError(
-            f"{path}.tool_name tool not registered: {tool_name}"
-        )
+        raise SolutionPlanningValidationError(f"{path}.tool_name tool not registered: {tool_name}")
     effect = _require_string(action, "effect", path)
     if effect not in _ACTION_EFFECT_VALUES:
         raise SolutionPlanningValidationError(f"{path}.effect is invalid")
@@ -999,124 +999,31 @@ def _validated_resource_ref_objects(
     return result
 
 
-def _validated_string_refs(
-    value: object,
-    allowed: set[str],
-    path: str,
-    label: str,
-) -> list[str]:
-    refs = _require_string_list(value, path)
-    for item in refs:
-        if item not in allowed:
-            raise SolutionPlanningValidationError(f"{label} reference does not exist: {item}")
-    return refs
-
-
-def _provider_summary(result: StructuredLLMResult) -> dict[str, object]:
-    return {
-        "provider": result.provider,
-        "model": result.model,
-        "requested_mode": result.requested_mode.value,
-        "actual_runtime": result.actual_runtime.value,
-        "input_tokens": result.input_tokens,
-        "output_tokens": result.output_tokens,
-        "total_tokens": result.total_tokens,
-        "latency_ms": result.latency_ms,
-        "fallback_reason": result.fallback_reason,
-        "structured_output_attempts": result.structured_output_attempts,
-        "provider_request_id": result.provider_request_id,
-        "safe_error_code": result.safe_error_code,
-    }
-
-
-def _require_schema_version(value: dict[str, object], path: str, expected: int) -> None:
-    schema_version = _require_int(value, "schema_version", path)
-    if schema_version != expected:
-        raise SolutionPlanningValidationError(f"{path}.schema_version must be {expected}")
-
-
-def _require_mapping(value: object, path: str) -> dict[str, object]:
-    if not isinstance(value, dict):
-        raise SolutionPlanningValidationError(f"{path} must be an object")
-    result: dict[str, object] = {}
-    for key, item in value.items():
-        if not isinstance(key, str):
-            raise SolutionPlanningValidationError(f"{path} keys must be strings")
-        result[key] = item
-    return result
-
-
-def _nullable_mapping(value: object, path: str) -> dict[str, object] | None:
-    if value is None:
-        return None
-    return _require_mapping(value, path)
-
-
-def _require_allowed_keys(
-    value: dict[str, object],
-    path: str,
-    *,
-    required: set[str],
-    optional: set[str],
-) -> None:
-    actual = set(value)
-    missing = required - actual
-    extra = actual - required - optional
-    if missing:
-        raise SolutionPlanningValidationError(
-            f"{path} is missing required fields: {sorted(missing)}"
-        )
-    if extra:
-        raise SolutionPlanningValidationError(f"{path} has unsupported fields: {sorted(extra)}")
-
-
-def _require_int(value: dict[str, object], field: str, path: str) -> int:
-    item = value[field]
-    if not isinstance(item, int) or isinstance(item, bool):
-        raise SolutionPlanningValidationError(f"{path}.{field} must be integer")
-    return item
-
-
-def _require_string(value: dict[str, object], field: str, path: str) -> str:
-    item = value[field]
-    if not isinstance(item, str):
-        raise SolutionPlanningValidationError(f"{path}.{field} must be string")
-    return item
-
-
-def _require_list(value: object, path: str) -> list[object]:
-    if not isinstance(value, list):
-        raise SolutionPlanningValidationError(f"{path} must be an array")
-    return value
-
-
-def _require_string_list(value: object, path: str) -> list[str]:
-    items = _require_list(value, path)
-    for index, item in enumerate(items):
-        if not isinstance(item, str):
-            raise SolutionPlanningValidationError(f"{path}[{index}] must be string")
-    return cast(list[str], items)
-
-
-def _optional_string_list(value: object) -> list[str]:
-    if value is None:
-        return []
-    items = _require_list(value, "$.clarification.list")
-    result: list[str] = []
-    for index, item in enumerate(items):
-        if not isinstance(item, str):
-            raise SolutionPlanningValidationError(
-                f"clarification list entry must be string: {index}"
-            )
-        result.append(item)
-    return result
-
-
-def _optional_option_list(value: object) -> list[dict[str, object]]:
-    if value is None:
-        return []
-    items = _require_list(value, "$.clarification.options")
-    return [_require_mapping(item, "$.clarification.options[]") for item in items]
+# Shared with the other agent workflow modules; see _schema_support module docstring.
+_require_mapping = partial(_schema.require_mapping, error_cls=SolutionPlanningValidationError)
+_nullable_mapping = partial(_schema.nullable_mapping, error_cls=SolutionPlanningValidationError)
+_require_allowed_keys = partial(
+    _schema.require_allowed_keys, error_cls=SolutionPlanningValidationError
+)
+_require_int = partial(_schema.require_int, error_cls=SolutionPlanningValidationError)
+_require_string = partial(_schema.require_string, error_cls=SolutionPlanningValidationError)
+_require_list = partial(_schema.require_list, error_cls=SolutionPlanningValidationError)
+_require_string_list = partial(
+    _schema.require_string_list, error_cls=SolutionPlanningValidationError
+)
+_require_schema_version = partial(
+    _schema.require_schema_version, error_cls=SolutionPlanningValidationError
+)
+_optional_string_list = partial(
+    _schema.optional_string_list, error_cls=SolutionPlanningValidationError
+)
+_optional_option_list = partial(
+    _schema.optional_option_list, error_cls=SolutionPlanningValidationError
+)
+_validated_string_refs = partial(
+    _schema.validated_string_refs, error_cls=SolutionPlanningValidationError
+)
+_provider_summary = _schema.provider_summary
 
 
 def _optional_string(value: object) -> str | None:
